+++ suzie_doctor/rootfs/app/suzie_doctor/__init__.py
@@ -1,5 +1,5 @@
 """Suzie Doctor App package."""
 
-APP_VERSION = "0.2.34-dev"
+APP_VERSION = "0.2.35-dev"
 PROTOCOL_PACK_VERSION = "0.1.4-dev"
 BRIDGE_VERSION = "0.2.4-dev"